Implement the delete product logic

+++ shopManagementTool/src/main/java/com/example/shopManagementTool/service/ProductService.java
@@ -29,14 +29,14 @@ public class ProductService {
     }
 
     public ResponseEntity<Product> validateAndSaveProduct(ProductDO productDO) {
-        logger.atDebug().log("validateAndSaveProduct in ProductService");
+        logger.atDebug().log("validateAndSaveProduct in ProductService.");
 
         if (containsIllegalsCharacters(productDO.getName())
                 && containsIllegalsCharacters(productDO.getDescription())
                 && (Double.isNaN(productDO.getPrice()) || productDO.getPrice() <= 0)
                 && productDO.getQuantity() <= 0) {
 
-            logger.atDebug().log("The new product to be saved contains invalid data");
+            logger.atDebug().log("The new product to be saved contains invalid data.");
 
             return new ResponseEntity<>(null, HttpStatus.NOT_ACCEPTABLE);
 
@@ -48,7 +48,7 @@ public class ProductService {
         }
     }
     public Product saveProduct(ProductDO productDO) {
-        logger.atDebug().log("checkAndSaveProduct in ProductService");
+        logger.atDebug().log("checkAndSaveProduct in ProductService.");
 
         try {
             Product product = productConverter.convertProductDOToBE(productDO);
@@ -68,8 +68,18 @@ public class ProductService {
         return updatedProduct;
     }
 
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public ResponseEntity<Product> deleteProduct(Long id) {
+        logger.atInfo().log("Delete the product.");
+        try {
+            if (productRepository.existsById(id)) {
+                logger.atDebug().addArgument(id).log("The product with id {} exist and can be deleted.");
+                productRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            logger.atError().addArgument(e).log("Unable to delete tge product. Reason: {}");
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>(null, HttpStatus.OK);
     }
 
     public boolean containsIllegalsCharacters(String toCheck) {
